requirements.txt
wfdb
numpy
pandas
sklearn
timm
transformers
--find-links https://download.pytorch.org/whl/cu116
torch 
torchvision
torchaudio
neptune-client
